disable erternal lib warnings

+++ sandbox/src/Render/Vertex.cpp
@@ -10,8 +10,8 @@ namespace sandbox
 
 	Vertex::Vertex(Vec2f Pos, Color Color)
 	{
-		data[0] = Pos.x;
-		data[1] = Pos.y;
+		data[0] = (float)Pos.x;
+		data[1] = (float)Pos.y;
 		data[2] = 0.f;
 		data[3] = (float)Color.r / 255;
 		data[4] = (float)Color.g / 255;
